Refactor after package rename
